server/requirements.txt
fastapi
uvicorn[standard]
monai
torch
pillow
numpy
matplotlib
